Add files via upload
How we build the project
First, we created a Player class to represent black and white pieces. Then we created Board class to
represent the state of the problem using a 2D array. And in the Board class, we also generated
basic_moves, capture_moves, valid_moves functions, and movement functions to represent the
transition model. In addition, we implemented Minimax, Alpha-beta pruning, H-Minimax with a
fixed depth cut off in Minimax class. Lastly, we created the game body in the main class.
How to run my project’s program
The program uses standard input and standard output to interact with the user. Follow the prompts
to choose your game setting. To move, our program accepts 4 digits(row1, col1, row2, col2). For
example, you can either enter “0110” or “0 1 1 0” to move from row:0 col:1 to row:1 col:0.
The program plays checkers on a 4x4 board or 8x8 board based on the user’s choice. For both
games, we use an adversarial state-space search approach to solve the problem. A player wins by
capturing all their opponent’s pieces and lose if they cannot make a legal move. The game will end
in a draw if there is no winner after 10 moves for a 4x4 game or 50 moves for an 8x8 game. They
can choose to play as Black(first) or White(second). They can choose their opponent type: random
AI, Minimax, Alpha-beta pruning, and H-Minimax with a fixed depth cut off. For both games, the
program has all 4 options. But we suggest using H-Minimax with a fixed depth cut off for 8x8 game
considering the search time.